- split RouteDisplayPage into smaller components - display textual directions and mode selector
